+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+### [2.1.4]
+### Fixed
+
+- Send buffering event to Conviva on seek/timeshift
+
+### [2.1.3]
+### Fixed
+
+- Fixed remaining unit tests
+
 ### [2.1.2]
 ### Fixed